StandardYoungTableaux

Overview

Member functions

StandardYoungTableaux.IPartition — Type

IPartition(partition)

Defines an integer partition.

Argument

partition: vector of positive integers in decreasing order

StandardYoungTableaux.StandardYoungTableau — Type

StandardYoungTableau(tableau)

Defines a standard Young tableau.

Argument

tableau: a vector of vectors of positive integers

StandardYoungTableaux.RS — Method

RS(sigma)

Pair of standard Young tableaux given by the Robinson-Schensted correspondence from a permutation.

Argument

sigma: a permutation given as a vector of integers

Example

RS([3, 4, 1, 2])

StandardYoungTableaux.SYT2YoungPath — Method

SYT2YoungPath(syt)

Path of the Young graph corresponding to a standard Young tableau.

Argument

syt: a standard Young tableau

Example

y = StandardYoungTableau([[1,3,4], [2]])
SYT2YoungPath(y)

StandardYoungTableaux.SYT2ballot — Method

SYT2ballot(syt)

Ballot sequence corresponding to a standard Young tableau.

Argument

syt: a standard Young tableau

StandardYoungTableaux.YoungPath2SYT — Method

YoungPath2SYT(path)

Standard Young tableau corresponding to a path of the Young graph.

Argument

path: a vector of integers partitions forming a path of the Young graph

Example

y = StandardYoungTableau([[1,3,4], [2]])
path = SYT2YoungPath(y)
YoungPath2SYT(path)

StandardYoungTableaux.allSYTx — Method

allSYTx(lambda)

All standard Young tableaux of a given shape.

Argument

lambda: an integer partition

Example

lambda = IPartition([3, 1])
allSYTx(lambda)

StandardYoungTableaux.ballot2SYT — Method

ballot2SYT(b)

Standard Young tableau corresponding to a ballot sequence.

Argument

b: a ballot sequence (vector of integers)

StandardYoungTableaux.countSYTx — Method

countSYTx(lambda)

Number of standard Young tableaux of a given shape.

Argument

lambda: an integer partition, the shape

Example

lambda = IPartition([4, 2])
countSYTx(lambda)
length(allSYTx(lambda))

StandardYoungTableaux.dualSYT — Method

dualSYT(syt)

Dual (conjugate) standard Young tableau.

Argument

syt: a standard Young tableau

StandardYoungTableaux.firstSYT — Method

firstSYT(lambda)

The 'first' standard Young tableau of a given shape.

Argument

lambda: an integer partition

StandardYoungTableaux.hooklengths — Method

hooklengths(lambda)

The hook lengths of a partition.

Argument

lambda: an integer partition

StandardYoungTableaux.hooks — Method

hooks(lambda)

The hooks of a partition.

Argument

lambda: an integer partition

StandardYoungTableaux.nextSYT — Method

nextSYT(syt)

The 'next' standard Young tableau.

Argument

syt: a standard Young tableau

Example

y = firstSYT(IPartition([4, 2]))
nextSYT(y)